synchronized::multiqueue moved to ordered::multiqueue
